Packages and DataGenerator class for tests
- New packages for units tests and their respective helpers
- DataGenerator class to generate data for testing purposes
